Abstração de funções js para utils
