+++ bi-front/src/components/OngPosts/Post.jsx
@@ -2,6 +2,7 @@ import PropTypes from "prop-types";
 import Button from "../Button/Button";
 import styles from "./post.module.css";
 import PostSelected from "./PostSelected";
+import { formatarData, calcularTempoRestante } from "../../utils/formatDate";
 
 export default function Post({
   selected,
@@ -59,39 +60,3 @@ Post.propTypes = {
   databaseState: PropTypes.array.isRequired,
   setDatabaseState: PropTypes.func.isRequired,
 };
-
-function calcularTempoRestante(dataExpiracao) {
-  const dataAtual = new Date();
-  const diferencaEmMilissegundos = dataExpiracao - dataAtual;
-  const umDia = 24 * 60 * 60 * 1000; // Milissegundos em um dia
-  const umaHora = 60 * 60 * 1000; // Milissegundos em uma hora
-
-  if (diferencaEmMilissegundos < 0) {
-    return "Post Expirado!"; //Se passou da data de expiração
-  }
-
-  const diasRestantes = Math.floor(diferencaEmMilissegundos / umDia);
-  const horasRestantes = Math.floor(
-    (diferencaEmMilissegundos % umDia) / umaHora
-  );
-
-  let resposta = null;
-  if (horasRestantes !== 0 && diasRestantes !== 0) {
-    resposta = `Tempo restante: ${diasRestantes} dias e ${horasRestantes} horas`;
-  } else if (horasRestantes === 0 && diasRestantes !== 0) {
-    resposta = `Tempo restante: ${diasRestantes} dias`;
-  } else if (horasRestantes !== 0 && diasRestantes === 0) {
-    resposta = `Tempo restante: ${horasRestantes} horas`;
-  } else {
-    resposta = "Menos de 1 hora restante";
-  }
-
-  return resposta;
-}
-
-function formatarData(data) {
-  const dia = String(data.getDate()).padStart(2, "0");
-  const mes = String(data.getMonth() + 1).padStart(2, "0"); // Janeiro é 0!
-  const ano = String(data.getFullYear()).slice(-2); // Pega os últimos 2 dígitos do ano
-  return `${dia}/${mes}/${ano}`;
-}
